Seperated the Traverser from the Output generation.
The Traverser now simply walks through a document. The OutputRules convert
the nodes into output html. The rules is a configurable options. By default
OutputRules will generate html close to the html5 that was parsed. Alternate
rule implementation (e.g., minify rules, pretty spacing rules) can be set
as the default or on an individual case.